Fixed syntax

+++ motortest.py
@@ -72,7 +72,7 @@ print("Big and small step running")
 
 GPIO.output(BIG_DIR, CW)
 GPIO.output(SMALL_DIR, CW)
-for i in range(REPS*SPIN*BIG_SPR)):
+for i in range(int(REPS*SPIN*BIG_SPR)):
     GPIO.output(BIG_STEP, GPIO.HIGH)
     time.sleep(b_delay)
     GPIO.output(BIG_STEP, GPIO.LOW)
